requirements.txt
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.3
scikit-learn==1.0.2
notebook==6.4.12
seaborn==0.11.2
pyyaml==6.0
scikit-survival==0.17.2
eli5==0.13.0
shap==0.41.0
xgboost==1.6.1
lifelines==0.27.1
xgbse==0.2.3
mrmr-selection==0.2.5
xgbse==0.2.3